+++ cmd/hidden_lake/applications/messenger/internal/config/editor_test.go
@@ -23,7 +23,7 @@ var (
 func (p *tsConfig) GetSettings() IConfigSettings     { return nil }
 func (p *tsConfig) GetLanguage() language.ILanguage  { return 0 }
 func (p *tsConfig) GetLogging() logger.ILogging      { return nil }
-func (p *tsConfig) GetShare() bool                   { return false }
+func (p *tsConfig) GetShareEnabled() bool            { return false }
 func (p *tsConfig) GetPseudonym() string             { return "" }
 func (p *tsConfig) GetAddress() IAddress             { return nil }
 func (p *tsConfig) GetNetworkKey() string            { return "" }