+++ cmd/hidden_lake/applications/messenger/pkg/settings/settings.go
@@ -31,7 +31,7 @@ const (
 )
 
 const (
-	CDefaultShare            = false
+	CDefaultShareEnabled     = false
 	CDefaultLanguage         = "ENG"
 	CDefaultMessagesCapacity = (2 << 10) // count
 )